% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LocationScaleRegression}
\alias{LocationScaleRegression}
\title{R6 class for location-scale regression models}
\description{
This model class assumes a normally distributed response variable with one
linear predictor for the mean and one for the standard deviation. The linear
predictors for the mean and the standard deviation are called \eqn{X\beta}
and \eqn{Z\gamma} respectively. The standard deviation uses a log link.
}
\examples{

## ------------------------------------------------
## Method `LocationScaleRegression$new`
## ------------------------------------------------

y <- rnorm(30)
LocationScaleRegression$new(y ~ 1)


## ------------------------------------------------
## Method `LocationScaleRegression$loglik`
## ------------------------------------------------

y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$loglik()


## ------------------------------------------------
## Method `LocationScaleRegression$grad`
## ------------------------------------------------

y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$grad()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameters}}{A named list of the \code{beta} and \code{gamma} parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocationScaleRegression$new()}}
\item \href{#method-loglik}{\code{LocationScaleRegression$loglik()}}
\item \href{#method-grad}{\code{LocationScaleRegression$grad()}}
\item \href{#method-clone}{\code{LocationScaleRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$new(
  mformula,
  sformula = ~1,
  data = environment(mformula),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mformula}}{A two-sided formula with the response variable on the
LHS and the predictor for the mean on the RHS.}

\item{\code{sformula}}{A one-sided formula with the predictor for the standard
deviation.}

\item{\code{data}}{A data frame (or list or environment) in which to evaluate
the formulas.}

\item{\code{...}}{Passed on to \code{\link[stats:model.matrix]{stats::model.matrix()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{LocationScaleRegression} object.
}

\subsection{Returns}{
A \code{LocationScaleRegression} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
LocationScaleRegression$new(y ~ 1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loglik"></a>}}
\if{latex}{\out{\hypertarget{method-loglik}{}}}
\subsection{Method \code{loglik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$loglik()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the log-likelihood of a \code{LocationScaleRegression} object at the
current parameter values.
}

\subsection{Returns}{
A single number.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$loglik()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grad"></a>}}
\if{latex}{\out{\hypertarget{method-grad}{}}}
\subsection{Method \code{grad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$grad(parameters = c("beta", "gamma"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{The names of the parameter blocks with respect to
which the gradient should be computed. Either \code{"beta"}
or \code{"gamma"} or both.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Returns the gradient of the log-likelihood of a
\code{LocationScaleRegression} object at the current parameter values.
}

\subsection{Returns}{
A named list of numeric vectors.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$grad()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
